services/loyalty-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
strawberry-graphql[fastapi]==0.236.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
aio-pika==9.4.1
pydantic==2.7.1
